Add initial version of the plugin
This add the NodeProxy, which currently only instantites the scene in
the path property in the editor and intentionally does nothing when run
in game. In short, it allows lazy loading of nodes when in game.

To use it simply put the path of the scene in NodeProxy and it will be
part of the scene tree in the editor but will not be in the scene tree
of the game. Example usage includes loading the player animation only
in the editor and when in game allows the developer to manually
instantite the player animation from the path property of the NodeProxy.

For future development reference only, a possible extension is add
the properties to allow and control lazy loading behavior in the game,
and add the properties to control loading behavior in the editor.